requirements.txt
dotenv
litellm>=1.74.3
pyyaml
pytest>=7.0.0
python-dotenv>=1.0.0
colorlog
langgraph==0.0.27
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0
arxiv==2.1.0
PyPDF2==3.0.0
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0